Made some changes to refactor the codes into a class in a package
