+++ text-ui-test/input.txt
@@ -12,4 +12,4 @@ list
 deadline cs assignment
 deadline return book /by Sunday
 list
- bye
+bye
